change to float

+++ header_files/canny.hpp
@@ -3,6 +3,6 @@
 #include "main.hpp"
 #include <cmath>
 #include <stdio.h>
-void noise_reduce(double**, double**);
-void intensity_gradient(double **);
+void noise_reduce(float**, float**);
+void intensity_gradient(float **);
 #endif
